Redesigned approach to discriminated unions
This approach is based on an empty interface for each discriminated
union. Then the schemas referenced by the union each have this interface
added to their classes. For deserialization, a custom converter may be
used to construct the correct type. For outgoing values, consumers can
simply construct the right type and set onto the property typed as the
interface.