Treat response sets (OpenApiResponses) as interfaces
This will make a cleaner experience for casting result types. Also,
switching to referring to these as ResponseSets instead of Responses
to avoid confusion with the plural of Response, which isn't the same
thing.

Also, changed ResponseBase in the SDK to OperationBase, and added
IOperationBase to have a base for the interface.

Note: Interfaces aren't yet working for common responses from the
components section, coming soon.