Rewrite schema property enricher to run after generation
Motivation
----------
The current approach to property enrichment is to enrich it before it is
added to the syntax tree. This has the disadvantage that there is a lot
of missing contextual information. For example, the SemanticModel can't
be used to get information about the type being referenced, so we can't
have different behaviors for value and reference types.

Modifications
-------------
Move the CSharpCompilation object creation to before syntax tree
generation and make it available on the GenerationContext. Also, add the
IServiceProvider to GenerationContext to make it more easily accessible.

Create a new DI system for delayed enrichers based around
IOpenApiSyntaxNodeEnricher and some extension methods. Apply these
enrichers after generation to mutate the existing syntax trees.

Drop schema property enrichers from the old enrichment system and change
them to be IOpenApiSyntaxNodeEnrichers.

In RequiredPropertyEnricher, make all ref-type properties nullable even
if they are required. This suppresses compiler warnings and ensures
consumers perform property null checks before assuming the server
returned valid data.

Results
-------
This more powerful enrichment concept is proved out. We can move forward
with rewriting all enrichment to run after generation.