Move base type registration to service, and drop feature system
Motivation
----------
The feature system doesn't make a lot of sense when we're running a DI
container per generation, rather than reusing the DI container, now that
we're exposing the IServiceProvider in GenerationContext. Also, the
feature system as written isn't thread-safe.

Modifications
-------------
Drop the feature system altogether, and instead just adding singleton
services to the DI container. Convert existing features to this
approach.