Add compilation enrichment support
Motivation
----------
Allow extensions to more easily mutate the entire compilation.

Modifications
-------------
Add ICompilationEnricher, and update OpenApi syntax tree enrichment to
be an ICompilationEnricher class.

Rework OpenApi syntax tree enrichers to use OpenApiEnrichmentContext,
which contains the OpenApi element, so that the CSharpCompilation can be
passed through the enrichment process. This is because the copy on
GenerationContext wouldn't be updated until after all enrichments are
complete, but we need an ongoing, mutated copy during enrichment.