Wireup method to deserialize the response body
Because reading the content and therefore deserialization is async, this
drops the Body property, instead providing a GetBodyAsync method to get
the body. It forwards the deserialization to the
ITypeSerializerRegistry.

I'm still not a huge fan of the way we're getting some of the
well-known type names from the SDK root namespace, it feels a bit hacky.
But we can refactor this later.