Simplify LocatedOpenApiElement parent handling
In most cases, we only need the immediate parent. We can reach the
parents higher up the tree via recursion. So we can improve perf by
dropping the array of parents and instead store only the immediate
parent on each element. For the cases where we need to recurse, an
extension method based on IEnumerable which can be short-circuited by
.First() does the trick.