Generate a separate response for each operation
If an operation uses a shared response class, implement a new response
that inherits from the response and adds the interface instead of adding
the interface to the shared response class. This will keep the interface
count from getting out of hand if the response is used many times. It
will also make the API surface a bit less confusing to consume since
every response for an operation will have a named class starting with
the operation name.