Improve the authentication infrastructure
Provide a system where authenticators can be registered in a class and
are automatically chosen from based on which schemes are supplied and
which schemes are configured on each operation.

This is done via a shared Authenticators class (which will get
properties in a future commit) combined with a SecuritySchemeSet
attribute placed on each request class.