Rework type names to YardarmTypeInfo
This provides a lot more details about a type, such as is it being
generated by an ITypeGenerator or is it just a reference to a
pre-existing type. This is going to help address issues in future
pull requests around how names are built and help with choosing the
correct names for schemas embedded within other non-generated types
such as parameters and headers.