Drop ElementTypeNameProvider
This is pretty redundant at this point, instead just use the type
generator registry directly and ask for the type name. The only purpose
this piece of middleware was serving was to deal with changing from
ILocatedApiElement to ILocatedApiElement<T>, and that is now better
handled directly on the type generator registry using reflection.