Add inherited request class for the selected media type
This also changes from an interface for each operation request to a
base class. This is another step towards having a request type for each
possible content type on a request. It also significantly cleans up how
schemas were being handled for request bodies, which was a bit of a
confusing mess. Now each request type can simply directly reference a
body schema instead of a special request body type which embeds the body
schema.