Ensure consistent header type names
Motivation
----------
There is a race condition surrounding header references with custom
schemas. If the reference in a response is processed before the header
component then incorrect type names are used.

Modifications
-------------
Add intelligence to the TypeGeneratorRegistry so it always handles
references to components as root elements, even if they are looked up
the first time using a reference to the component.

Rework the ResponseTypeGenerator to get type names from the header type
generator instead of the schema type generator.

Fixes #23