Improve command line file path handling
Motivation
----------
The current commandline for output of files and packages is a bit
cumbersome, requiring lots of file paths.

Modifications
-------------
Allow the main output parameters to be a directory (indicated by a
trailing slash). This then outputs all of the files to that directory,
named based on the assembly name. For NuGet packages, include the
version number in the file name.

If the main output parameter is a file, then by default output the
other supporting files to the same directory with the same name, just
changing the extension. This is overridden if a particular path is
provided for that file instead.

Add additional parameters which can suppress the output of unwanted
files.

Fixes #5