Move AdditionalProperties to an enricher (#27)
Motivation
----------
Since enrichers run after all classes are generated, we can look at the
base type (if any) and determine information about its
AdditionalProperties.

Modifications
-------------
Create the enricher and move most of the logic from
ObjectSchemaGenerator to the enricher. Add additional logic to either
specify the shadowing of the base property with a `new` keyword, or
suppress generation altogether.

Tweak the run order of the JsonAddtionalPropertiesEnricher so it runs
afterwards.

Results
-------
If the base class already has AdditionalProperties and the type is the
same (very commonly dynamic), then the child class gets no new
declaration. If the base class uses a different type, then shadow the
declaration in the base class.

Fixes #26