Add MultipartPropertyAttribute for property name override
Motivation
----------
Much like JsonProperty for Newtonsoft.Json, we need some attribute to
indicate the attribute name (with correct casing) to use when
serializing a schema using multipart/form-data encoding.

Modifications
-------------
Add a new attribute and apply it using an enricher to all schema
properties.

Drop the JsonProperty attribute which was being incorrectly applied to
the body property of Request classes.