Allow additional properties on nested classes
Motivation
----------
If a schema has additionalProperties set to true and also has a nested
schema, and the outer class is an allOf inherited from a parent schema,
we fail with the semantic model.

Modifications
-------------
When enriching a class with AdditionalProperties, check the base type
on the semantic model using the original node rather than the node which
was mutated by enriching the child class.