Add support for multiple categories of type generators (#46)
Motivation
----------
For System.Text.Json support we want to also generate types and be able
to access their names from ITypeGeneratorRegistry. This is primarily to
create custom converters for discriminated schemas.

Modifications
-------------
Expand the TypeGeneratorRegistry system to work on an additional generic
type key. This type is not used/instantiated, and is purely present to
act as a DI key for lookup.

For compatibility, leave the existing interfaces which accept a single
type as inherited from the new interfaces using the new
PrimaryGeneratorCategory used for our existing generators.

Results
-------
Extensions may now register their own generator types using alternative
TGeneratorCategory key classes.