Add discriminator support to System.Text.Json
Motivation
----------
System.Text.Json doesn't support polymorphic deserialization out of the
box, but discriminators are a common pattern in OpenAPI specs.

Modifications
-------------
Implement a code generator to create a custom deserializer for each
schema with a discriminator, and register this on the base interface
with an attribute.

Add a unit test project to test some of the details.