Convert dynamic properties to JsonElement
Motivation
----------
System.Text.Json can't handle dynamic properties, so use JsonElement
instead.

Modifications
-------------
Add an enricher which performs the conversion.

Note: This could have unwanted side effects in the future if we're
mixing JSON and non-JSON content types using the same schema. But we
don't do that now so we'll cross that bridge when we come to it.