Support strongly-typed writes of undiscriminated oneOf schemas
Motivation
----------
There are cases where oneOf may be used on a request body schema but
not on the response, but which don't have a discriminator. This case is
supportable, though deserialization is not.

Modifications
-------------
In OneOfSchemaGenerator generate an interface even when there is no
discriminator property. Also make it public and add the generator to the
interface so we can identify it.

Add extensions to help collect all distinct schemas from an
OpenApiDocument, recursing through all layers.

Generate JsonConverters in System.Text.Json for all schemas which are
handled by the OneOfSchemaGenerator, even if they don't have a
discriminator property. Also ensure we get unique names for the
converters.

In that case where we don't have a discriminator property throw an
exception on read in the converter.