Include more complete preprocessor symbols for resources
Motivation
----------
We want to start supporting targets other than netstandard2.0.

Modifications
-------------
Add a CurrentTargetFramework property to GenerationContext, always set
to netstandard2.0 for now.

Update ResourceSyntaxTreeGenerator to add standard multi-targeting
preprocessor symbols for most of the frameworks we care about, based on
the CurrentTargetFramework.

Move some constants to be more accessible.

Results
-------
No significant difference for now, but a step closer to multi-target.