Handle AdditionalProperties inheritance regardless of execution order
Motivation
----------
There is logic in place that checks base classes of schema models for
existing instances of the AdditionalProperties property. However, the
enricher is checking for its own properties, meaning if an inherited
class is enriched before the base class then it won't see the property
and assume reimplementation is required. This results in warnings and
serialization errors.

Modifications
-------------
Instead of looking at actual properties on the base class, instead look
at the additional properties settings on the base class' schema. This
will always be present even if the base class hasn't been enriched.