Simplify constructors for OperationResponse
Motivation
----------
Improve the ability to mock operation responses for unit tests which
consume a mock API.

Modifications
-------------
Add additional constructors to OperationResponse and its descendants
that don't require a TypeSerializerRegistry. Add a static instance of
the registry which is used as the fallback.

Change ParseHeaders to take the headers as a parameter rather than pull
them from the HttpResponseMessage. This will allow the addition of
other constructor overloads which receive headers elsewhere in future
commits.

Relates to #77