Provide SerializationData to serializers (framework only)
Motivation
----------
Serializers, such as the multi-part serializer, require additional data
to function. This provides the framework of a method to provide this
data, though more work is required.

Modifications
-------------
Include a read-only, static SerializationData property on media type
specific request classes. This always returns null for now. Forward
this on to the serializer.

Call IRequestMemberGenerators for media type specific request classes,
and update them all to be more intelligent and recognize if they should
generate a member or not based on the mediaType parameter.

All IRequestMemberGenerators to return a variable number of members,
including zero members.