Remove the ISerializationData concept
Motivation
----------
After more careful review, having ISerializationData passed externally
to the ITypeSerializer instances was adding a lot of logical overhead in
the request classes. Instead, we can simply rely on the type T passed to
the serializer to acquire then necessary information.

Modifications
-------------
Remove ISerializationData, revert the recent changes to add it to the
request classes, and make MultipartFormSerializationData internal.