Send multipart requests
Motivation
----------
Multipart requests should be sent with correct form fields.

Modifications
-------------
Replace MultipartEncoding with MultipartPropertyInfo, and make it and
MultipartFormDataSerializationData generic.

Initialize those types using lambda expressions to prevent the need for
reflection.

Add a BinaryStreamSerializer to send binary multipart data.

Drop the MultipartPropertyAttribute, it is now unnecessary and handled
by the MultipartFormDataSerializationData class.

Results
-------
`multipart/form-data` requests are now supported with
`application/octet-stream`, `text/plain`, and `application/json` content
types. The first listed content type from the spec is always assumed.
This will need to be improved on in the future.

We no longer apply the wasteful `[MultipartProperty("...")]` attributes
to every property on every schema, we simply add the info to the
multipart serialization data only on requests that involve multipart
encoding.