Use BinaryStream serializer as a fallback for Stream and byte[]
Motivation
----------
Anytime we have a Stream or byte[] schema we should fallback to using
the BinarySerializer, even if the Content-Type isn't recognized.

Modifications
-------------
Add `Type` based key indexing to `TypeSerializerRegistry` with handling
for inherited types.

Register the `BinarySerializer` for `Stream` (and therefore all types
inherited from `Stream` and `byte[]`.

When finding a serializer, fallback to schema type lookup if media type
lookup doesn't find a match.