Allow manual setting of content type on multipart requests (#94)
Motivation
----------
It is generally necessary to provide a content type other than the
default for a multipart request when dealing with binary data.

Modifications
-------------
For multipart requests, inherit from the schema and add a content type
property for each schema property. Supply a lambda to get this property
to the `MultipartPropertyInfo`. When left as null, fallback to the
first content type in the array as before.