Improve trimmability
Motivation
----------
Eventually we'd like the generated SDKs to be trimmable.

Modifications
-------------
Move away from reflection to find methods and use `.GetMethodInfo()` on
lambda expressions instead, where possible.

Add `DynamicallyAccessedMembers` annotations where feasible when
targeting .NET 6.

Results
-------
The generated assemblies are still not trimmable due to the use of
`TypeDescriptor.GetConverter` as well as non-trimmable JSON serializers.
However, we've taken a step in the right direction and improved
trimming in cases where the consumer is trying to manually control
trimming.