Move doc/settings to YardarmGenerator constructor (#106)
Motivation
----------
Now that we have multiple methods on YardarmGenerator, it makes more
sense to pass in dependencies and build the service provider once in the
constructor instead of on every method call. This will allow reuse of
generated values across multiple calls if the YardarmGenerator instance
is retained.

Modifications
-------------
Move OpenApiDocument and YardarmGenerationSettings to the constructor,
and build the IServiceProvider in the constructor.

Drop those parameters from the public methods.

Remove YardarmGenerationSettings from several private methods and use
the instance field instead.

Add a method to get the package spec. This will likely be useful for
calculating the dependency graph from a MSBuild project.