Add collect-package-references command
Motivation
----------
Provide an extension point that outputs package references in a format
which can be read by an MSBuild task. This will allow NuGet restore
from an MSBuild project to know the correct packages to restore.

Modifications
-------------
Add the command and DTOs for serializing JSON information on the console
to be received by MSBuild.