Generate deterministic source paths
Motivation
----------
Produce symbols that have deterministic source paths like other C#
projects when built on CI servers.

Modifications
-------------
Ensure the embedded source files have a "/_0" prefix. This way the
source path will be the same when the same OpenAPI spec is built,
regardless of the file path to the OpenAPI spec. This is detected as
deterministic by other tooling such as NuGet Package Explorer.