Output files to intermediate locations until generate succeeds
Motivation
----------
The current generate process overwrites files with zero length files in
the case of a failure. This both destroys potentially valid previous
generate results and causes confusion for incremental builds because
the timestamp on the files is updated.

Modifications
-------------
Write all outputs to temporary files when run via the command line and
copy to the final destination only on success.