Split interface and class tag generators
Motivation
----------
There is not currently a means for extensions to determine the name of
the API implementation class, only the API interface.

Modifications
-------------
- Split the generation of the implementation out to a new category of
  generator and new implementation.
- Move some common logic into a shared abstract class.
- Move the TagComparer out to a shared public instance.

Results
-------
Extensions may use the TypeGeneratorRegistry for OpenApiTag with the
TagImplementationCategory to get the TypeInfo, etc, for the
implementation class.

They may also use the TagComparer instance to get a distinct list of
tags.