Normalize embedded source file paths (#138)
Motivation
----------
The current paths are convoluted and difficult to read in error
messages and stack traces. Also, since tag elements now generate two
files only one gets included in the PDB and linked to both the
interface and the class. Finally, the source paths are not deterministic
between Windows and Linux builds.

Modifications
-------------
- For element-based source files, use the namespace and type name to
  determine the file name. This should still guarantee uniqueness.
- For resource files and the assembly info file, use forward slashes
  instead of back slashes for OS consistency.