Provide more global defaults for APIs with DI
Motivation
----------
It's very common to have defaults that apply to all APIs within the SDK,
such as handlers to add to the pipeline or a customer configuration for
the primary handler.

Modifications
-------------
Allow these to be registered globally and then let settings on specific
APIs take precedence.

- Primary HTTP handler
- Additional HTTP handlers (these are added before specific API handlers
  so they are an outer wrapper in the pipeline)
- Handler lifetime
- Log header redaction

Refactor the current approach to configuring the HttpClient to work in
tandem with these new settings.

Results
-------
Many common global settings can be applied once instead of requiring
them to be applied to each API:

- Circuit breaker patterns with Polly that will break the circuit for
  the entire SDK
- Custom logging middleware
- Log redaction
- Custom SSL configuration on the primary handler
- etc