Allow configuration of authenticators in DI
Motivation
----------
Having the default authenticators live in the DI container isn't very
helpful if there isn't also a way to configure them.

Modifications
-------------
Add `ConfigureAuthenticators` method to `IApiBuilder` with an additional
overload which receives the `IServiceProvider` to allow configuration to
be acquired from the DI container.

Add a similar overload to `ConfigureBaseAddress` which accepts an
`IServiceProvider` for dynamic configuration.