Add support for "partial-time" string formats
Motivation
----------
This is the standard method OpenAPI specs encode time-only properties.

Modifications
-------------
Treat "partial-time" as a `TimeSpan` and add support to the `LiteralSerializer`. Both JSON libraries have out-of-the-box support.

Also treat "full-date" as the equivalent of "date" per the spec.

Results
-------
"partial-time" is now fully supported as a `TimeSpan`.

Note: This does not address "time" or "time-only" formats, which also encode the time zone.  This is more difficult to represent in .NET, we may need to use `DateTimeOffset`

Closes #204