Add support for cancellation during deserialization
Motivation
----------
Now that we support streaming response bodies to the deserializer it's
more likely that we could encounter a cancellation while deserializing.
We should respect the request for cancellation where possible.

Modifications
-------------
Add an overload of `DeserializeAsync` to `ITypeSerializer` and
`TypeSerializerRegistryExtensions` that accepts a cancellation token.
For modern runtimes add a default implementation for the new overload
in `ITypeSerializer`.

Use the cancellation token parameter wherever possible. For legacy
frameworks that don't accept the cancellation token when requesting the
response body from `HttpResponseContent` at least verify the token
before making the call.

Results
-------
Cancellation is now supported for most cases, including all of the most
common cases, when deserializing.

**BREAKING CHANGE**

Custom implementations of `ITypeSerializer` in an extension must
implement both overloads if they support generating SDKs targeting .NET
Standard 2.0.

This may also affect SDK consumers if they use their own custom
`ITypeSerializer` implementations and are using a .NET Standard 2.0 SDK.
This could occur if the consumer is targeting .NET Framework, the
consumer is an intermediate NuGet package targeting .NET Standard 2.0,
or any case where the SDK is only distributed as a .NET Standard 2.0
package without a secondary modern runtime such as .NET 6 included in
the NuGet package. It is recommended that any SDK distribution that
includes .NET Standard 2.0 also include .NET 6 assembly at a minimum.