Generate correct code for parameter components (#243)
The current code for parameter components fails because it can't generate a source file path for embedding. Also, it results in duplicate schema generation if there are references to schemas from the parameters and can have name conflicts between these schemas.

Instead, generate parameter classes only in cases where a nested schema needs to be generated and nest the schema inside that class. This also addresses the source file name problem since there is a type name to use as the basis for the file name.

Relates to #239