Generate properties for parameters included on path items
Currently, properties are only generated for parameters listed
explicitly on each operation (except for simple `string` properties for
route parameters). The OpenAPI spec allows for parameters to be defined
at the path level and then overridden for each operation.

This implements that behavior, handling query and header parameters
defined at the path level and using the correct schema for route
parameters defined at the path level.