Enable Server GC
Motivation
----------
Since we are highly parallelized and doing a lot of SyntaxNode heap
allocations, for multicore machines we can get fewer GC pauses by using
server GC. This is a good default for this project.

Modifications
-------------
Enable Server GC for the Yardarm command line. Note that a single core
machine will ignore this setting and use Workstation GC.