Allow the command line app to roll forward to .NET 9 runtime
Motivation
----------
It's possible for a system running a build to have only the .NET 9 SDK
installed, without the .NET 8 runtime. In this case, setting environment
variables to force a roll forward to .NET 9 is required.

Modifications
-------------
Set the roll forward configuration to Major in the command line app's
runtimeconfig.json.

Results
-------
SDK-style builds will now automatically roll forward to the .NET 9
runtime to run Yardarm if the .NET 8 runtime is unavailable. For best
compatibility, the .NET 8 runtime is still prefered if present.