Add DateOnly and TimeOnly literal converts
Motivation
----------
We'd like to add support for DateOnly and TimeOnly when targeting
modern .NET and using System.Text.Json. When we do, we'll need support
for header and query string literals as well.

Modifications
-------------
- Add DateOnlyLiteralConverter and TimeOnlyLiteralConverter, mirroring
  the serialization formats used by System.Text.Json
- Register on LiteralConverterRegistry

Results
-------
No changes for now unless an extension is used to change schema
generation to use DateOnly and TimeOnly.