Add support for DateOnly and TimeOnly
Motivation
----------
For "date", "partial-date", "time", and "partial-time" string formats,
DateOnly and TimeOnly are a better fit when available in .NET 6 or
later.

Modifications
-------------
- Add a dependency on Microsoft.Extensions.Options
- Create a new internal option to control date/time handling, and
  initialize based on command-line properties and target frameworks
- Use the new option to choose schema types
- Treat "time" format the same as "partial-time"

Breaking Change
---------------
When building targeting only .NET 6 or later, the default type used
for "date", "partial-date", and "partial-time" is now DateOnly or
TimeOnly instead of DateTime or TimeSpan. This may cause backward
compatibility issues in generated SDKs. Use the command line switch
`-p LegacyDateTimeHandling=true` to restore the previous behavior. This
is not necessary if targeting .NET Standard 2.0.