Add an option to relax discriminator rules
Motivation
----------
Some APIs may have more flexible response types that cause
deserialization errors as new discriminator values are added.

Modifications
-------------
Add a new property which may be passed with a value of
`UnknownDiscrimiatorHandling=ReturnNull` to opt out of JsonException and
instead return null when an unknown or missing discriminator is
encountered.