Allow setting the HTTP version (#310)
Motivation
----------
By default, .NET will only use HTTP/1.1. If an API supports HTTP/2.0 or 3.0 we need an easy way for the generated SDK to enable this support on compatible runtimes.

Modifications
-------------
Accept two new parameters, `DefaultHttpVersion` and `DefaultHttpVersionPolicy`, which set these values. The second is ignored on downlevel runtimes. If not set, they will default to the values used by default in HttpRequestMessage for the target runtime.

Fixes #300